.github/scripts/requirements.txt
pyyaml==6.0.2
jinja2==3.1.4